==> Code/00_Data_Wrangling.ipynb <==
%matplotlib inline

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt 
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler

# Training set
X_train = pd.read_csv(r".\data\UCI HAR Dataset\UCI HAR Dataset\train\X_train.txt", header=None, delim_whitespace=True)
y_train = pd.read_csv(r".\data\UCI HAR Dataset\UCI HAR Dataset\train\y_train.txt", header=None, delim_whitespace=True)

# Test set
X_test = pd.read_csv(r".\data\UCI HAR Dataset\UCI HAR Dataset\test\X_test.txt", header=None, delim_whitespace=True)
y_test = pd.read_csv(r".\data\UCI HAR Dataset\UCI HAR Dataset\test\y_test.txt", header=None, delim_whitespace=True)

# Variables labels
features = pd.read_csv(r".\data\UCI HAR Dataset\UCI HAR Dataset\features.txt", header=None, sep = " ")

# subjects names
subject_test = pd.read_csv(r".\data\UCI HAR Dataset\UCI HAR Dataset\test\subject_test.txt", header=None)
subject_train = pd.read_csv(r".\data\UCI HAR Dataset\UCI HAR Dataset\train\subject_train.txt", header=None)

#### RINOMINA VARIABILI

# Lista con nomi delle variabili
features.rename({1: 'Features'}, axis=1, inplace=True)
features_list = list(features['Features'].values)

# Rinomina variabili per X_train
for i in range(len(features_list)):
    X_train.rename({i: str(features_list[i])}, axis=1, inplace=True)

# Rinomina variabili per X_test
for i in range(len(features_list)):
    X_test.rename({i: str(features_list[i])}, axis=1, inplace=True)
    
# Rinomina variabile target y_train
y_train.rename({0: 'Label'}, axis=1, inplace=True)

# Rinomina variabile target y_test
y_test.rename({0: 'Label'}, axis=1, inplace=True)

#### AGGIUNTA DI COLONNA SUBJECT

In totale ci sono 30 subjects. 21 per il training set e 9 per il test set.

# aggiungere subject al train
X_train['Subject'] = subject_train[0]

# aggiungere subject al test
X_test['Subject'] = subject_test[0]

# distinct values di subject_test
X_test['Subject'].value_counts()

# distinct values di subject_train
X_train['Subject'].value_counts()

The "Label" stands for:
- 1: "WALKING" 
- 2: "WALKING_UPSTAIRS"
- 3: "WALKING_DOWNSTAIRS"
- 4: "SITTING"
- 5: "STANDING"
- 6: "LAYING"

### X_train

X_train.head()

# nel caso si volesse fare un file csv
#X_train.to_csv(r'.\data\csv_data\X_train.csv', index=False)

### y_train

y_train.head()

# nel caso si volesse fare un file csv
#y_train.to_csv(r'.\data\csv_data\y_train.csv', index=False)

### X_test

X_test.head()

# nel caso si volesse fare un file csv
#X_test.to_csv(r'.\data\csv_data\X_test.csv', index=False)

### y_test

y_test.head()

# nel caso si volesse fare un file csv
#y_test.to_csv(r'.\data\csv_data\y_test.csv', index=False)

## Aggiungere variabile Target al Train e al Test sets

# Copia dataset con X_train
train = X_train.copy()
# Aggiunta della variabile target (y_train) al train set X_train
train["Label"] = y_train["Label"]

# Copia dataset con X_test
test = X_test.copy()
# Aggiunta della variabile target (y_test) al test set X_test
test["Label"] = y_test["Label"]

### Train set 

train.head()

train.info()

# nel caso si volesse fare un file csv
#train.to_csv(r'.\data\csv_data\train.csv', index=False)

### Test set

test.head()

test.info()

# nel caso si volesse fare un file csv
#test.to_csv(r'.\data\csv_data\test.csv', index=False)

##  Whole dataset

# Dataframe di tutto il dataset (+ una variabile per indicare se appartiene al train o al test set)
df = pd.concat([test.assign(ind="test"), train.assign(ind="train")])

# Dataframe di tutto il dataset (senza possibilità di ridividerlo tra train e test set)
# df = pd.concat([train, test])

df.head()

df.info()

# nel caso si volesse fare un file csv
#df.to_csv(r'.\data\csv_data\df.csv', index=False)

# nel caso facciamo modifiche a tutto il dataset (df) e poi vogliamo ridividerlo tra train e test
# test, train = df[df["ind"].eq("test")], df[df["ind"].eq("train")]

## Are there any missing values?

# missing values in df
df.isnull().values.any()

# numero di NaN in df
df.isnull().sum().sum()

## Is the dataset balanced?

# Target variable value counts in the whole dataset (df)
df['Label'].value_counts()

# Target variable value counts in the train set (y_train)
y_train.value_counts()

# Target variable value counts in the test set (y_test)
y_test.value_counts()
