==> Code/11_Classification_SupportVectorMachines.ipynb <==
# Support Vector Machines

%matplotlib inline

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.metrics import accuracy_score, f1_score, classification_report
from sklearn.metrics import roc_curve, auc, roc_auc_score
import seaborn as sns
sns.set_style(style="whitegrid")

from collections import defaultdict
import warnings
warnings.filterwarnings('ignore')

# datasets
X_train = pd.read_csv('.\data\csv_data\X_TR_151_SENZA_OUTLIERS.csv')
X_test = pd.read_csv('.\data\csv_data\X_TST_151_SENZA_OUTLIERS.csv')
y_train = pd.read_csv('.\data\csv_data\Y_TR_151_SENZA_OUTLIERS.csv')
y_test = pd.read_csv('.\data\csv_data\y_test.csv')

# Target variable from integer to categorical
y_train['Label'] = pd.Categorical(y_train["Label"])
y_test['Label'] = pd.Categorical(y_test["Label"])

from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
scaler.fit(X_train)
X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)

# LINEAR SVM

from sklearn.svm import LinearSVC

clf = LinearSVC(C=0.01, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = LinearSVC(C=0.1, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = LinearSVC(C=1.0, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)
print('')
print(classification_report(y_test, y_pred))

clf = LinearSVC(C=10.0, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = LinearSVC(C=100.0, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

# NON LINEAR SVM

from sklearn.svm import SVC

## kernel = 'poly'

clf = SVC(gamma='auto', kernel = 'poly', C=0.01, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = SVC(gamma='auto', kernel = 'poly', C=0.1, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = SVC(gamma='auto', kernel='poly', C=1.0, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = SVC(gamma='auto', kernel='poly', C=10.0, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = SVC(gamma='auto', kernel='poly', C=100.0, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)
print('')
print(classification_report(y_test, y_pred))

## kernel = 'rbf'

clf = SVC(gamma='auto', kernel='rbf', C=0.01, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = SVC(gamma='auto', kernel='rbf', C=0.1, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = SVC(gamma='auto', kernel='rbf', C=1.0, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = SVC(gamma='auto', kernel='rbf', C=10.0, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = SVC(gamma='auto', kernel='rbf', C=100.0, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

## kernel = 'sigmoid'

clf = SVC(gamma='auto', kernel='sigmoid', C=0.01, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = SVC(gamma='auto', kernel='sigmoid', C=0.1, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = SVC(gamma='auto', kernel='sigmoid', C=1.0, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = SVC(gamma='auto', kernel='sigmoid', C=10.0, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))

clf = SVC(gamma='auto', kernel='sigmoid', C=100.0, random_state=42)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(classification_report(y_test, y_pred))
