==> Code/17_Regression_GradientBoosting.ipynb <==
# Gradient Boosting for Regression

%matplotlib inline

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.metrics import accuracy_score, f1_score, classification_report
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
from sklearn.metrics import roc_curve, auc, roc_auc_score

import warnings
warnings.filterwarnings('ignore')

# datasets
train = pd.read_csv('.\data\csv_data\X_TR_151_SENZA_OUTLIERS.csv')
test = pd.read_csv('.\data\csv_data\X_TST_151_SENZA_OUTLIERS.csv')

# creo X_train & y_train come per la regressione lineare
X_train = train.drop('tBodyAcc-entropy()-X', axis=1)
y_train = pd.DataFrame(train['tBodyAcc-entropy()-X'])

X_train.shape

X_train.info()

y_train.shape

y_train.info()

# divido allo stesso modo il test
X_test = test.drop('tBodyAcc-entropy()-X', axis=1)
y_test = pd.DataFrame(test['tBodyAcc-entropy()-X'])

**Gradient Boosting Regressor**

from sklearn.ensemble import GradientBoostingRegressor

>class sklearn.ensemble.GradientBoostingRegressor(*, loss='squared_error', learning_rate=0.1, n_estimators=100, subsample=1.0, criterion='friedman_mse', min_samples_split=2, min_samples_leaf=1, min_weight_fraction_leaf=0.0, max_depth=3, min_impurity_decrease=0.0, init=None, random_state=None, max_features=None, alpha=0.9, verbose=0, max_leaf_nodes=None, warm_start=False, validation_fraction=0.1, n_iter_no_change=None, tol=0.0001, ccp_alpha=0.0)

reg = GradientBoostingRegressor(random_state=42) # max_depth=3
reg.fit(X_train, y_train)
y_pred = reg.predict(X_test)

print('R2: %.5f' % r2_score(y_test, y_pred))
print('MSE: %.5f' % mean_squared_error(y_test, y_pred))
print('MAE: %.5f' % mean_absolute_error(y_test, y_pred))

reg = GradientBoostingRegressor(max_depth=5, random_state=42)
reg.fit(X_train, y_train)
y_pred = reg.predict(X_test)

print('R2: %.5f' % r2_score(y_test, y_pred))
print('MSE: %.5f' % mean_squared_error(y_test, y_pred))
print('MAE: %.5f' % mean_absolute_error(y_test, y_pred))
