==> Code/24_TimeSeries_Classification_Rocket_MiniRocket.ipynb <==
# Time Series Classification

%matplotlib inline

import numpy as np 
import pandas as pd
from sktime.utils.plotting import plot_series
from sktime.datatypes._panel._convert import from_3d_numpy_to_nested
import matplotlib.pyplot as plt

## TRAIN DATA

# TRAINING SET files

# total acc
total_acc_x_train = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\train\Inertial Signals\total_acc_x_train.txt', header = None, delim_whitespace=True).values
total_acc_y_train = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\train\Inertial Signals\total_acc_y_train.txt', header = None, delim_whitespace=True).values
total_acc_z_train = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\train\Inertial Signals\total_acc_z_train.txt', header = None, delim_whitespace=True).values

# body acc
body_acc_x_train = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\train\Inertial Signals\body_acc_x_train.txt', header = None, delim_whitespace=True).values
body_acc_y_train = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\train\Inertial Signals\body_acc_y_train.txt', header = None, delim_whitespace=True).values
body_acc_z_train = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\train\Inertial Signals\body_acc_z_train.txt', header = None, delim_whitespace=True).values

# body gyro
body_gyro_x_train = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\train\Inertial Signals\body_gyro_x_train.txt', header = None, delim_whitespace=True).values
body_gyro_y_train = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\train\Inertial Signals\body_gyro_y_train.txt', header = None, delim_whitespace=True).values
body_gyro_z_train = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\train\Inertial Signals\body_gyro_z_train.txt', header = None, delim_whitespace=True).values

# y
y_train = pd.read_csv(r'.\data\csv_data\y_train.csv')

# subjects
subjects = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\train\subject_train.txt', header = None) 

# CREATING NESTED ARRAYS

# stack the time stamps together as 1D
# total acc
total_acc_x_train = np.dstack(total_acc_x_train)
total_acc_y_train = np.dstack(total_acc_y_train)
total_acc_z_train = np.dstack(total_acc_z_train)

# body acc
body_acc_x_train = np.dstack(body_acc_x_train)
body_acc_y_train = np.dstack(body_acc_y_train)
body_acc_z_train = np.dstack(body_acc_z_train)

# body gyro
body_gyro_x_train = np.dstack(body_gyro_x_train)
body_gyro_y_train = np.dstack(body_gyro_y_train)
body_gyro_z_train = np.dstack(body_gyro_z_train)


# move the axis 
# total acc
total_acc_x_train = np.moveaxis(total_acc_x_train, -1, 0)
total_acc_y_train = np.moveaxis(total_acc_y_train, -1, 0)
total_acc_z_train = np.moveaxis(total_acc_z_train, -1, 0)

# body acc
body_acc_x_train = np.moveaxis(body_acc_x_train, -1, 0)
body_acc_y_train = np.moveaxis(body_acc_y_train, -1, 0)
body_acc_z_train = np.moveaxis(body_acc_z_train, -1, 0)

# body gyro
body_gyro_x_train = np.moveaxis(body_gyro_x_train, -1, 0)
body_gyro_y_train = np.moveaxis(body_gyro_y_train, -1, 0)
body_gyro_z_train = np.moveaxis(body_gyro_z_train, -1, 0)


# transform the 3D numpy arrays to nested ones
total_acc_x_train = from_3d_numpy_to_nested(total_acc_x_train)
total_acc_y_train = from_3d_numpy_to_nested(total_acc_y_train)
total_acc_z_train = from_3d_numpy_to_nested(total_acc_z_train)

# body acc
body_acc_x_train = from_3d_numpy_to_nested(body_acc_x_train)
body_acc_y_train = from_3d_numpy_to_nested(body_acc_y_train)
body_acc_z_train = from_3d_numpy_to_nested(body_acc_z_train)

# body gyro
body_gyro_x_train = from_3d_numpy_to_nested(body_gyro_x_train)
body_gyro_y_train = from_3d_numpy_to_nested(body_gyro_y_train)
body_gyro_z_train = from_3d_numpy_to_nested(body_gyro_z_train)

# X TRAIN DATASET
variables = [total_acc_x_train, total_acc_y_train, total_acc_z_train, 
             body_acc_x_train, body_acc_y_train, body_acc_z_train,
             body_gyro_x_train, body_gyro_y_train, body_gyro_z_train]
variables_names = ['total_acc_x', 'total_acc_y', 'total_acc_z',
                   'body_acc_x', 'body_acc_y', 'body_acc_z',
                   'body_gyro_x', 'body_gyro_y', 'body_gyro_z']
X_train = pd.concat(variables, axis = 1) #keys = variables_names)
X_train.columns = variables_names

X_train.head()

X_train.shape

## TEST DATA

# TEST SET files

# total acc
total_acc_x_test = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\test\Inertial Signals\total_acc_x_test.txt', header = None, delim_whitespace=True).values
total_acc_y_test = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\test\Inertial Signals\total_acc_y_test.txt', header = None, delim_whitespace=True).values
total_acc_z_test = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\test\Inertial Signals\total_acc_z_test.txt', header = None, delim_whitespace=True).values

# body acc
body_acc_x_test = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\test\Inertial Signals\body_acc_x_test.txt', header = None, delim_whitespace=True).values
body_acc_y_test = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\test\Inertial Signals\body_acc_y_test.txt', header = None, delim_whitespace=True).values
body_acc_z_test = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\test\Inertial Signals\body_acc_z_test.txt', header = None, delim_whitespace=True).values

# body gyro
body_gyro_x_test = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\test\Inertial Signals\body_gyro_x_test.txt', header = None, delim_whitespace=True).values
body_gyro_y_test = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\test\Inertial Signals\body_gyro_y_test.txt', header = None, delim_whitespace=True).values
body_gyro_z_test = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\test\Inertial Signals\body_gyro_z_test.txt', header = None, delim_whitespace=True).values

# y
y_test = pd.read_csv(r'.\data\csv_data\y_test.csv')

# subjects
subjects = pd.read_csv(r'.\data\UCI HAR Dataset\UCI HAR Dataset\test\subject_test.txt', header = None) 

# CREATING NESTED ARRAYS

# stack the time stamps together as 1D
# total acc
total_acc_x_test = np.dstack(total_acc_x_test)
total_acc_y_test = np.dstack(total_acc_y_test)
total_acc_z_test = np.dstack(total_acc_z_test)

# body acc
body_acc_x_test = np.dstack(body_acc_x_test)
body_acc_y_test = np.dstack(body_acc_y_test)
body_acc_z_test = np.dstack(body_acc_z_test)

# body gyro
body_gyro_x_test = np.dstack(body_gyro_x_test)
body_gyro_y_test = np.dstack(body_gyro_y_test)
body_gyro_z_test = np.dstack(body_gyro_z_test)


# move the axis 
# total acc
total_acc_x_test = np.moveaxis(total_acc_x_test, -1, 0)
total_acc_y_test = np.moveaxis(total_acc_y_test, -1, 0)
total_acc_z_test = np.moveaxis(total_acc_z_test, -1, 0)

# body acc
body_acc_x_test = np.moveaxis(body_acc_x_test, -1, 0)
body_acc_y_test = np.moveaxis(body_acc_y_test, -1, 0)
body_acc_z_test = np.moveaxis(body_acc_z_test, -1, 0)

# body gyro
body_gyro_x_test = np.moveaxis(body_gyro_x_test, -1, 0)
body_gyro_y_test = np.moveaxis(body_gyro_y_test, -1, 0)
body_gyro_z_test = np.moveaxis(body_gyro_z_test, -1, 0)


# transform the 3D numpy arrays to nested ones
total_acc_x_test = from_3d_numpy_to_nested(total_acc_x_test)
total_acc_y_test = from_3d_numpy_to_nested(total_acc_y_test)
total_acc_z_test = from_3d_numpy_to_nested(total_acc_z_test)

# body acc
body_acc_x_test = from_3d_numpy_to_nested(body_acc_x_test)
body_acc_y_test = from_3d_numpy_to_nested(body_acc_y_test)
body_acc_z_test = from_3d_numpy_to_nested(body_acc_z_test)

# body gyro
body_gyro_x_test = from_3d_numpy_to_nested(body_gyro_x_test)
body_gyro_y_test = from_3d_numpy_to_nested(body_gyro_y_test)
body_gyro_z_test = from_3d_numpy_to_nested(body_gyro_z_test)

# X TEST DATASET
variables = [total_acc_x_test, total_acc_y_test, total_acc_z_test, 
             body_acc_x_test, body_acc_y_test, body_acc_z_test,
             body_gyro_x_test, body_gyro_y_test, body_gyro_z_test]
variables_names = ['total_acc_x', 'total_acc_y', 'total_acc_z',
                   'body_acc_x', 'body_acc_y', 'body_acc_z',
                   'body_gyro_x', 'body_gyro_y', 'body_gyro_z']
X_test = pd.concat(variables, axis = 1) #keys = variables_names)
X_test.columns = variables_names

X_test.head()

X_test.shape

# trasformare ys in arrays
y_train = y_train['Label'].values
y_test = y_test['Label'].values

## ROCKET

from sklearn.linear_model import RidgeClassifierCV
from sktime.transformations.panel.rocket import Rocket

%%time
rocket = Rocket()  # by default, MiniRocket uses ~10,000 kernels
rocket.fit(X_train)
X_train_transform = rocket.transform(X_train)
clf = RidgeClassifierCV(alphas=np.logspace(-3, 3, 10), normalize=True)
clf.fit(X_train_transform, y_train)

from sklearn.metrics import accuracy_score, f1_score, classification_report
from sklearn.metrics import roc_curve, auc, roc_auc_score

X_test_transform = rocket.transform(X_test)
y_pred = clf.predict(X_test_transform)
print(classification_report(y_test, y_pred))

## MINI-ROCKET

from sktime.transformations.panel.rocket import MiniRocketMultivariate

%%time
minirocket = MiniRocketMultivariate()  # by default, MiniRocket uses ~10,000 kernels
minirocket.fit(X_train)
X_train_transform = minirocket.transform(X_train)
clf = RidgeClassifierCV(alphas=np.logspace(-3, 3, 10), normalize=True)
clf.fit(X_train_transform, y_train)

X_test_transform = minirocket.transform(X_test)
y_pred = clf.predict(X_test_transform)
print(classification_report(y_test, y_pred))
